Add files via upload
I have implemented undo. It was a long and arduous journey, but it is done. It appears to function correctly even under difficult QA tests.
In order to do so, I've changed a number of backend variables, including adding some basic regular expressions and making a QList<String *> variable. this has been an interesting challenge, but I am glad to say that the core functionality of the calculator is fully implemented.

For my next trick, I plan to connect keyboard buttons to the buttons, and to allow for copy pasting numbers from other sources.

I've also included some screenshots of the calculator for reference.