Add files via upload
I've changed a lot of stuff, but because of those changes, the calculator now accepts keyboard buttons. 
This includes changing the way CalcButton interfaces with the Calculator functions (I no longer use singals and slots, but instead buttons now store a pointer to the calc and call a public function do_event() when they're pressed). Implementing do_event() also allowed me to remove event_char_from_button(). I've also rearranged all the member functions in the header, and edited many of the comments and some of the formatting of the other functions.

I also include some updated screenshots (I'm deleting the old ones) and some photos of the paper I used to plan out the basic aesthetics and functionality of the calculator before I started programming.